// LaNube/src/app/panel-admin/interfaces/asistencia.interface.ts
export interface EstadoAlumnoInterface {
    id:          number;
    idAlumno:    number;
    bath1:       string;
    bath2:       string;
    bath3:       string;
    eat1:        string;
    eat2:        string;
    eat3:        string;
    horaEntrada: string;
    horaSalida:  string;
    asistencia:  boolean;
    fecha:       Date;
}